--- lib/widgets/user_transactions.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:my_expense_app/widgets/transaction_list.dart';
-
-import '../models/transaction.dart';
-import 'new_transaction.dart';
-
-class UserTransations extends StatefulWidget {
-
-  @override
-  State<UserTransations> createState() => _UserTransationsState();
-}
-
-class _UserTransationsState extends State<UserTransations> {
-  final List<Transaction> transactions = [
-    Transaction(
-      id: 't1',
-      title: 'New Shoes',
-      amount: 69.99,
-      date: DateTime.now(),
-    ),
-    Transaction(
-      id: 't2',
-      title: 'Weekly Groceries',
-      amount: 16.53,
-      date: DateTime.now(),
-    ),
-  ];
-
-  void _addTransaction(String title, double amount) {
-    final newTx = Transaction(
-      title: title,
-      amount: amount,
-      date: DateTime.now(),
-      id: DateTime.now().toString(),
-    );
-    setState(() {
-      transactions.add(newTx);
-    });
-  }
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        NewTransaction(_addTransaction),
-        TransactionList(transactions),
-      ],
-    );
-  }
-}
